schwarz-precond: hoist lsmr to crate root, drop solve namespace
With CG and GMRES gone, the solve module wrapped a single solver and
its name redundant with the only function inside (solve::lsmr::lsmr).
Move src/solve/lsmr.rs and src/solve/lsmr/ up one level to src/lsmr.rs
and src/lsmr/, fold the dot/vec_norm helpers from solve.rs into lsmr.rs
as pub(crate), and re-export lsmr/mlsmr/LsmrResult/LsmrStopReason from
the crate root alongside SchwarzPreconditioner.
