+++ CHANGELOG.md
@@ -28,7 +28,8 @@ and this project follows [Semantic Versioning](https://semver.org/).
 ### Removed
 
 - **BREAKING:** `schwarz_precond::solve::{cg, gmres}` removed. Migrate to
-  `solve::lsmr::mlsmr`
+  the unpreconditioned `lsmr` and preconditioned `mlsmr` re-exported from
+  the crate root.
 - **BREAKING:** `schwarz_precond::schwarz::multiplicative` removed, along
   with `MultiplicativeSchwarzPreconditioner`, `ResidualUpdater`, and
   `OperatorResidualUpdater`. Use `SchwarzPreconditioner` (additive) instead.
